+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add a direct keyboard shortcut to copy a Markdown link to the current page.
 - Copy the text/html version of a link in addition when copying a Markdown link.
+- Implement a help modal.
 - Fix Markdown escaping.
 - Fix unreliable background saving to Raindrop.
 - Ignore invalid timestamp strings.